Fix long levels

+++ src/player.c
@@ -132,7 +132,7 @@ void player_main() {
         }
 
         // Check if the level complete cutscene should start
-        s64 player_x_limit = ((curr_level_width << 4) - 0x98) << (SUBPIXEL_BITS);
+        s64 player_x_limit = (s64)((curr_level_width << 4) - 0x98) << (SUBPIXEL_BITS);
         if (curr_player.player_x > player_x_limit) {
             complete_cutscene = TRUE;
